Work around: prevent leaking more and more watchers
For now, when a container requests a connection, a new provider is created per container.
If that container goes away, the provider will still be there, watching the backend store for changes.
This causes a memory leak, since they are never cleaned up.
For now, this commit improves it by not creating watchers and fetching the values for every new TCP connection.
Providers will still pile up though, so this still needs to be fixed.